Added transpose_rows and corrected main functions
